better errors

+++ resources/views/pages/show_companies.blade.php
@@ -33,13 +33,15 @@ var Tr = React.createClass({
                 <td>{this.props.data.commercial_name}</td>
                 <td className="center"> <label className="input-control checkbox"> <input type="checkbox" checked={Boolean(JSON.parse(this.props.data.active))} /> <span className="check"></span> </label> </td> 
                 <td className="center"><img className="company_logo" src={this.props.data.logo} /></td>
-                <td> 
-                    <a href={"/companies/"+this.props.data.company_id+"/edit"} className="button success">{{trans('general.modify')}}</a>
-                    &nbsp;
-                    <button className="button warning delete_item" data-type="companies" data-id={this.props.data.company_id}>{{trans('general.disable')}}</button>
-                    &nbsp;
-                    <a href={"/companies/"+this.props.data.company_id+"/users"} className="button" data-type="companies" data-id={this.props.data.company_id}>{{trans_choice('general.menu.users', 2)}}</a>
-                </td>
+                @if(Auth::user()->can('edit-companies'))
+                    <td> 
+                        <a href={"/companies/"+this.props.data.company_id+"/edit"} className="button success">{{trans('general.modify')}}</a>
+                        &nbsp;
+                        <button className="button warning delete_item" data-type="companies" data-id={this.props.data.company_id}>{{trans('general.disable')}}</button>
+                        &nbsp;
+                        <a href={"/companies/"+this.props.data.company_id+"/users"} className="button" data-type="companies" data-id={this.props.data.company_id}>{{trans_choice('general.menu.users', 2)}}</a>
+                    </td>
+                @endif
 
             </tr>
 
@@ -63,7 +65,9 @@ var CompanyTable = React.createClass({
                         <th> {{ trans('general.forms.commercial_name')}} </th>
                         <th> {{ trans('general.active')}} </th>
                         <th>logo</th>
-                        <th> {{ trans_choice('general.actions',2)}} </th>
+                        @if(Auth::user()->can('edit-companies'))
+                            <th> {{ trans_choice('general.actions',2)}} </th>
+                        @endif
                     </tr>
                 </thead>
                 <tbody>
