+++ resources/views/pages/show_departments.blade.php
@@ -5,7 +5,7 @@
 @section('content')
 
                 
-<h2> {{ trans_choice('general.menu.departments', 2) }} </h2>
+<h2> {{ trans_choice('general.menu.departments', 2) }} <a href="/departments/create" class="button success"> {{ trans('general.forms.add_new') }} </a></h2>
 <div id="table"></div>
 <hr>
 <button class="button success">{{ trans('general.forms.add_new') }}</button>
@@ -27,12 +27,14 @@ var Tr = React.createClass({
                 <td>{this.props.data.name}</td>
                 <td>{this.props.data.company_name}</td>
                 <td className="center"> <label className="input-control checkbox"> <input type="checkbox" checked={Boolean(JSON.parse(this.props.data.active))} /> <span className="check"></span> </label> </td> 
-                <td> 
-                    <a href={"/departments/"+this.props.data.department_id+"/edit"} className="button success">{{trans('general.modify')}}</a>
-                    &nbsp;
-                    <button className="button warning delete_item" data-type="departments" data-id={this.props.data.department_id}>{{trans('general.disable')}}</button>
-                    
-                </td>
+                @if(Auth::user()->can('edit-departments'))
+                    <td> 
+                        <a href={"/departments/"+this.props.data.department_id+"/edit"} className="button success">{{trans('general.modify')}}</a>
+                        &nbsp;
+                        <button className="button warning delete_item" data-type="departments" data-id={this.props.data.department_id}>{{trans('general.disable')}}</button>
+                        
+                    </td>
+                @endif
 
             </tr>
 
@@ -47,9 +49,13 @@ var CompanyTable = React.createClass({
             data: this.props.data
         };
     },
+    componentDidMount: function() {
+    
+        $('#datatable').DataTable();
+    },
     render: function() {
         return (
-            <table className="table striped hovered cell-hovered border bordered">
+            <table id="datatable" className="table striped hovered cell-hovered border bordered">
                 <thead>
                     <tr>
 
@@ -58,7 +64,9 @@ var CompanyTable = React.createClass({
                         <th> {{ trans('general.forms.name')}} </th>
                         <th>{{ trans_choice('general.menu.companies', 1) }}</th>
                         <th> {{ trans('general.active')}} </th>
-                        <th> {{ trans_choice('general.actions',2)}} </th>
+                        @if(Auth::user()->can('edit-departments'))
+                            <th> {{ trans_choice('general.actions',2)}} </th>
+                        @endif
                     </tr>
                 </thead>
                 <tbody>
